feat: unary ops
